<attachment> title text disappears when dragging https://bugs.webkit.org/show_bug.cgi?id=142263 <rdar://problem/20035515>
Reviewed by Sam Weinig.

* rendering/RenderThemeMac.mm:
(WebCore::RenderThemeMac::paintAttachment):
We use NSColors with CoreText, which requires that Cocoa knows the current
CGContext out-of-band (through the NSGraphicsContext currentContext mechanism).
Otherwise, sometimes (especially when dragging), it will try to apply the
text foreground color to a null context, complain to the system log, and fail
to render the text.


Identifier: 160325@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@181008 268f45cc-cd09-0410-ab3c-d52691b4dbfc