Rename MediaStreamSource to RealtimeMediaSource https://bugs.webkit.org/show_bug.cgi?id=142330
Reviewed by Eric Carlson.

MediaStreamSource is a misleading name, it's not the source of a
MediaStream.

The chain from high level to low level is: MediaStream -
MediaStreamTrack - MediaStreamTrackPrivate - RealTimeMediaSource
and a MediaStream can contains several MediaStreamTrack objects of
different types.

MediaStreamSourceStates was also renamed to
RealtimeMediaSourceStates, and MediaStreamSourceCapabilities to
RealtimeMediaSourceCapabilities, for consistency.

Identifier: 160436@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@181152 268f45cc-cd09-0410-ab3c-d52691b4dbfc